move material to linalg
